move FederatedProviderButton from primitives directory to ./Authenticator/common
