+++ packages/react-native/src/Authenticator/common/FederatedProviderButton/index.ts
