+++ packages/react-native/src/Authenticator/common/FederatedProviderButton/styles.ts
