+++ packages/react-native/src/Authenticator/common/index.ts
@@ -1,0 +1,1 @@
+export * from './FederatedProviderButton';
